Proof of concept for WordPress image generation
This was tested with WordPress Core's Docker setup, and WP-CLI.

The changes also include removing AVIF from being generated, and a size 85, which is close, but not quite, the core default.
Testing with a more focused set around the current JPEG defaults would be helpful.

TODO:
Make it easier to set this up for more local environments/paths, including by splitting out constants more.
Consider adding support for environment variables and/or arguments to set necessary paths.